[SW] Cleanup code
- Move state machine to seperate framework folder
- Remove deprecated files
- Move profiler tool into main.go
